+++ src/services/__mocks__/queueService.js
@@ -1,22 +1,28 @@
 // Manual mock for queueService - prevents Redis connections during tests
 
-const mockQueue = {
-  add: jest.fn().mockResolvedValue({ id: 'mock-job-id' }),
+// Mock queue service for testing
+const reportParsingQueue = {
   process: jest.fn(),
+  add: jest.fn().mockResolvedValue({ id: 'mock-job-id' }),
   on: jest.fn(),
-  close: jest.fn().mockResolvedValue(),
-  removeRepeatable: jest.fn().mockResolvedValue()
+  close: jest.fn().mockResolvedValue()
 };
 
-const addJob = jest.fn().mockResolvedValue(undefined);
+const telegramScrapingQueue = {
+  process: jest.fn(),
+  add: jest.fn().mockResolvedValue({ id: 'mock-scraping-job-id' }),
+  removeRepeatable: jest.fn().mockResolvedValue(),
+  on: jest.fn(),
+  close: jest.fn().mockResolvedValue()
+};
 
-const cleanup = jest.fn().mockResolvedValue(undefined);
-
-// Methods for Telegram scraping
-const triggerTelegramScraping = jest.fn().mockResolvedValue({
-  jobId: 'mock-scraping-job-id',
-  status: 'queued'
-});
+const reportProcessingQueue = {
+  process: jest.fn(),
+  add: jest.fn().mockResolvedValue({ id: 'mock-processing-job-id' }),
+  removeRepeatable: jest.fn().mockResolvedValue(),
+  on: jest.fn(),
+  close: jest.fn().mockResolvedValue()
+};
 
 const startTelegramScraping = jest.fn().mockResolvedValue({
   success: true,
@@ -28,11 +34,6 @@ const stopTelegramScraping = jest.fn().mockResolvedValue({
   message: 'Telegram scraping stopped'
 });
 
-const triggerManualScraping = jest.fn().mockResolvedValue({
-  id: 'mock-manual-scraping-job-id'
-});
-
-// Methods for batch report processing
 const startBatchReportProcessing = jest.fn().mockResolvedValue({
   success: true,
   message: 'Batch report processing started'
@@ -43,16 +44,24 @@ const stopBatchReportProcessing = jest.fn().mockResolvedValue({
   message: 'Batch report processing stopped'
 });
 
+const triggerManualScraping = jest.fn().mockResolvedValue({
+  id: 'mock-manual-job-id',
+  success: true
+});
+
+const addJob = jest.fn().mockResolvedValue({ id: 'mock-job-id' });
+
+const cleanup = jest.fn().mockResolvedValue();
+
 module.exports = {
-  addJob,
-  reportParsingQueue: mockQueue,
-  telegramScrapingQueue: mockQueue,
-  reportProcessingQueue: mockQueue,
-  cleanup,
-  triggerTelegramScraping,
+  reportParsingQueue,
+  telegramScrapingQueue,
+  reportProcessingQueue,
   startTelegramScraping,
   stopTelegramScraping,
   startBatchReportProcessing,
   stopBatchReportProcessing,
-  triggerManualScraping
+  triggerManualScraping,
+  addJob,
+  cleanup
 };
